# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Engineering Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: repositoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed, rolled_back]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Record a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDeployment
      tags: [deployments]
      summary: Update deployment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  # ─── Pull Requests ────────────────────────────────────────────────────────
  /pull-requests:
    get:
      operationId: listPullRequests
      tags: [pull-requests]
      summary: List pull requests
      parameters:
        - name: repositoryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
            enum: [open, merged, closed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of pull requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PullRequest"
    post:
      operationId: createPullRequest
      tags: [pull-requests]
      summary: Record a pull request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}:
    get:
      operationId: getPullRequest
      tags: [pull-requests]
      summary: Get a pull request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pull request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePullRequest
      tags: [pull-requests]
      summary: Update a pull request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"

  # ─── Incidents ────────────────────────────────────────────────────────────
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
            enum: [p1, p2, p3, p4]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Record an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  # ─── Event Ingestion ──────────────────────────────────────────────────────
  /events:
    post:
      operationId: ingestEvent
      tags: [events]
      summary: Ingest a raw activity event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    get:
      operationId: listEvents
      tags: [events]
      summary: List recent events
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

  # ─── Analytics / Metrics ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Teams ─────────────────────────────────────────────────────────────
...
    Deployment:
      type: object
      required: [id, repositoryId, environment, status, deployedAt]
      properties:
        id:
          type: integer
        repositoryId:
          type: integer
        teamId:
          type: ["integer", "null"]
        environment:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, rolled_back]
        commitSha:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        leadTimeSeconds:
          type: ["integer", "null"]
        deployedAt:
          type: string
          format: date-time
        finishedAt:
          type: ["string", "null"]
    DeploymentInput:
      type: object
      required: [repositoryId, environment, status]
      properties:
        repositoryId:
          type: integer
        teamId:
          type: integer
        environment:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, rolled_back]
        commitSha:
          type: string
        version:
          type: string
        leadTimeSeconds:
          type: integer
        deployedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    DeploymentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, success, failed, rolled_back]
        leadTimeSeconds:
          type: integer
        finishedAt:
          type: string
          format: date-time

    # ── Pull Requests ─────────────────────────────────────────────────────
    PullRequest:
      type: object
      required: [id, repositoryId, title, state, openedAt]
      properties:
        id:
          type: integer
        repositoryId:
          type: integer
        teamId:
          type: ["integer", "null"]
        externalId:
          type: ["integer", "null"]
        title:
          type: string
        authorLogin:
          type: ["string", "null"]
        state:
          type: string
          enum: [open, merged, closed]
        additions:
          type: ["integer", "null"]
        deletions:
          type: ["integer", "null"]
        changedFiles:
          type: ["integer", "null"]
        cycleTimeSeconds:
          type: ["integer", "null"]
        openedAt:
          type: string
          format: date-time
        mergedAt:
          type: ["string", "null"]
        closedAt:
          type: ["string", "null"]
    PullRequestInput:
      type: object
      required: [repositoryId, title, state, openedAt]
      properties:
        repositoryId:
          type: integer
        teamId:
          type: integer
        externalId:
          type: integer
        title:
          type: string
        authorLogin:
          type: string
        state:
          type: string
          enum: [open, merged, closed]
        additions:
          type: integer
        deletions:
          type: integer
        changedFiles:
          type: integer
        cycleTimeSeconds:
          type: integer
        openedAt:
          type: string
          format: date-time
        mergedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    PullRequestUpdate:
      type: object
      properties:
        state:
          type: string
          enum: [open, merged, closed]
        additions:
          type: integer
        deletions:
          type: integer
        changedFiles:
          type: integer
        cycleTimeSeconds:
          type: integer
        mergedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time

    # ── Incidents ─────────────────────────────────────────────────────────
    Incident:
      type: object
      required: [id, title, severity, status, openedAt]
      properties:
        id:
          type: integer
        teamId:
          type: ["integer", "null"]
        title:
          type: string
        severity:
          type: string
          enum: [p1, p2, p3, p4]
        status:
          type: string
          enum: [open, resolved]
        recoveryTimeSeconds:
          type: ["integer", "null"]
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
    IncidentInput:
      type: object
      required: [title, severity, status, openedAt]
      properties:
        teamId:
          type: integer
        title:
          type: string
        severity:
          type: string
          enum: [p1, p2, p3, p4]
        status:
          type: string
          enum: [open, resolved]
        recoveryTimeSeconds:
          type: integer
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    IncidentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, resolved]
        severity:
          type: string
          enum: [p1, p2, p3, p4]
        recoveryTimeSeconds:
          type: integer
        resolvedAt:
          type: string
          format: date-time

    # ── Events ────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, type, occurredAt]
      properties:
        id:
          type: integer
        teamId:
          type: ["integer", "null"]
        repositoryId:
          type: ["integer", "null"]
        type:
          type: string
        actorLogin:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [type]
      properties:
        teamId:
          type: integer
        repositoryId:
          type: integer
        type:
          type: string
          minLength: 1
        actorLogin:
          type: string
        metadata:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time

    # ── Analytics ─────────────────────────────────────────────────────────
...