# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * Engineering Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentInputStatus } from './deploymentInputStatus';

export interface DeploymentInput {
  repositoryId: number;
  teamId?: number;
  environment: string;
  status: DeploymentInputStatus;
  commitSha?: string;
  version?: string;
  leadTimeSeconds?: number;
  deployedAt?: Date;
  finishedAt?: Date;
}